Fix stuck hover states on tab close buttons
Use the |bounds| property instead of the |frame| property when updating tracking rectangles and deciding whether the button has the mouse over it.  We used to do this from the parent views context, but moved the code into the button itself and this broke.

Check the image state when the mouseDown event loop finishes.  We can get into a weird state when the tabs resize and the close button is placed under the cursor.  We update the image to highlight it, but if we click/hold the mouse button without moving the mouse, there are no mouseMove events and thus no mouseEntered, and thus no mouseExited (just assuming that's what is happening).

BUG=31279
TEST=Open a bunch of tabs, Click/hold on the last close button, then let go. Wit
hout moving the mouse at all, click/hold on the next button that moves under the
 mouse. Drag mouse away from tab strip area and let go. It should unhighlight it
self.

Review URL: http://codereview.chromium.org/524016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@35678 4ff67af0-8c30-449e-8e8b-ad334ec8d88c