Fix a bug of unnecessary "Confirm form resubmission" dialog.
If a URL is requested by the POST method and then the page is
redirected to another URL, NavigationEntry::has_post_data_ is not
cleared correctly with the prior code. So "Confirm form resubmission"
dialog is shown when a user asks reloading the redirected page.


TEST=Add a unit test
BUG=21245

Review URL: http://codereview.chromium.org/524057

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@35770 4ff67af0-8c30-449e-8e8b-ad334ec8d88c