Fix mis-sized download shelf close button
In an earlier checkin I had removed the images from the IB files that used HoverCloseButtons, relying on them to be set programmatically.  It seems that this causes a flaky sizing problem when loading from a NIB, which only manifested in the DownloadShelf for some reason.

Also change DownloadShelf.xib so that the close button is a child of the main view, not of the GTMWidthTweaker, which seems to eat redraw messages.

BUG=31785
TEST=Try out the download shelf a few times.  You need to quit the whole app a few times to ensure it doesn't come back.

Review URL: http://codereview.chromium.org/540001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@35959 4ff67af0-8c30-449e-8e8b-ad334ec8d88c