Rename CloseIdleConnections -> CloseCurrentConnections.
This method is (and was) just for debugging; while it was closing
idle connections, that was not aggressive enough; the benchmark needs
to close all connections, not just the idle ones.

To ensure connections are abandoned, create a new pool; leaving the old pool
to languish as any pending sockets die.

BUG=none
TEST=none


Review URL: http://codereview.chromium.org/549031

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@36208 4ff67af0-8c30-449e-8e8b-ad334ec8d88c