Another try at landing this. It caused several tests to fails previously. There was a crasher in some Windows code in the CLD that has been removed by the port of the CLD code to Linux and Mac. That should hopefully make everything work now.
Enabling language detection on page load.
A memory error has been fixed in the CLD library in the meantime.
This should hopefully fixes the crashers in the reliability tests.
Note that this version is actually simpler than the original review since the detection is now performed in the renderer. (So the CLD code runs sandboxed.)

Original review:
http://codereview.chromium.org/492024/show

BUG=30662
TEST=Run the unit-tests.
TBR=brettw

Review URL: http://codereview.chromium.org/552049

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@36589 4ff67af0-8c30-449e-8e8b-ad334ec8d88c