Fix crash in tab's close button.
If you click and hold on a close button that is being animated closed, it will crash if you release it after the tab is completely closed.  The button is being freed when the |super mouseDown| call is in it's tracking loop, so we need to hold on to it with a scoped_nsobject until we're done with it.  The checkImageState call is called unnecessarily, but it shouldn't hurt anything.

BUG=28220
TEST=Click and hold on the close button of a tab that is animating closed.  Release it a couple of seconds after the tab is fully hidden.  To help test, you can change the kAnimationDuration in tab_strip_controller.mm to a higher value.

Review URL: http://codereview.chromium.org/555018

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@36769 4ff67af0-8c30-449e-8e8b-ad334ec8d88c