Step 1 in Implementing/Prototyping App Panels.
In this first step we

1) Create a new Browser::Type::TYPE_APP_PANEL which is observed in various places
2) Create an AppPanelBrowserFrame which is created in BrowserFrameWin when the TYPE_APP_PANEL is observed. AppPanelBrowserFrame draws itself per glen's mocks and will ultimately behave substantially different from regular browser windows.
3) Create a temporary command switch called --app-launch-as-panel which can be used until we implement actual app launching.

Note that there is still work to be done "bit-twiddling" to match glen's mocks and to implement the "inactive" look for these windows.

Steps 2 through N may include: different sizing behavior (based either on the content size or an api call), different min/maximize behavior, docking to the os taskbar, auto order & placement of app panels.

A screen capture of the current implementation is attached to the bug.

BUG=32361

Review URL: http://codereview.chromium.org/553143

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@38107 4ff67af0-8c30-449e-8e8b-ad334ec8d88c