Revert 38323 - Add IPv6 probing support, and disable IPv6 resolution when it is useless
I've added minimal probing to check if IPv6 is at all possible, and when
it is not, then we disable IPv6 resolution.

I've also added histograms and A/B test support to evaluate the impact of
this change.

r=wtc,eroman
Review URL: http://codereview.chromium.org/579010

TBR=jar@chromium.org
Review URL: http://codereview.chromium.org/582011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@38327 4ff67af0-8c30-449e-8e8b-ad334ec8d88c