Reverting change: broke unit tests on chrome OS.
Revert 39143 - This change will include preferences as part of the usual startup sequence of the PSS that already includes bookmarks.  This is a temporary solution until we have a proper component to manage the startup and shutdown of multiple data types.

Review URL: http://codereview.chromium.org/601037

TBR=skrul@chromium.org
Review URL: http://codereview.chromium.org/611006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@39155 4ff67af0-8c30-449e-8e8b-ad334ec8d88c