Drop duplicate SBPrefix from full-hash elements.
SafeBrowsingDatabaseBloom stores a integer prefix column and a blob
hash column, where the first four bytes of the hash are the prefix.
This isn't necessary, so getting rid of the duplication rather than
propagating it into the future.

[Previously this was left alone because it wasn't too harmful because
the code could always drop the duplication from the file format.  A
future change for SafeBrowsingStoreFile will write an array of hashes,
making the duplication unavoidable.]

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/651057

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@39512 4ff67af0-8c30-449e-8e8b-ad334ec8d88c