Move data type start/stop management into DataTypeManager This change introduces a new interface/class called DataTypeManager whose job is to choreograph the start up and shut down of all registered data types.  It starts each data type serially, waiting for each data type to finish starting before starting the next one.  If anything goes wrong on startup, all data types are stopped.
Note that this change also simplifies the ProfileSyncServiceStartupTest as many of the cases it tested for are now the responsibility of the DataTypeManagerImpl (and these are thoroughly tested in its unit test).

I also killed off the TestProfileSyncFactory in the ProfileSyncServiceTest since it was lame and could be done with the mock.

BUG=36506

Review URL: http://codereview.chromium.org/650175

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@39964 4ff67af0-8c30-449e-8e8b-ad334ec8d88c