AutoFill Mac preference for enabling Address Book "me" card usage.
Adds a boolean preference to specify whether to use the Mac user's Address Book "me" card as a source of information for AutoFill addresses.  This CL adds the Mac UI and the preference setup only.  It does not cover the functionality for fetching the actual Address Book data.  The changes to AutoFillDialog.xib add the checkbox UI and bind the checkbox to the |auxiliaryEnabled| property in the controller class.  The |addressSection_| outlet is set to the spacer following the new checkbox.  Unit tests are included for the addition of the preference at the mac dialog controller level.

BUG=36496
TEST=PersonalDataManagerTest, AutoFillDialogControllerTest

Review URL: http://codereview.chromium.org/668171

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@40803 4ff67af0-8c30-449e-8e8b-ad334ec8d88c