Mac: Improve window look from "completely broken" to "pretty broken" when RI is enabled.
From http://developer.apple.com/mac/library/DOCUMENTATION/UserExperience/Conceptual/HiDPIOverview/HiDPISupport/HiDPISupport.html:

"Because of scaling, the coordinates of the window frame and its top-level view (the frame view) are not always the same. [...] These methods provide the means to transform geometry to and from a base coordinate space that is pixel-aligned with the backing store into which the view is being drawn."

Stuff is still pretty broken with this patch (it used to be better), but it's a low-hanging fruit.

BUG=19476
TEST=Open QuartzDebug, set scale factor to e.g. 1.5. Reopen Chrome. Tabs should be visible.

Review URL: http://codereview.chromium.org/893001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@41336 4ff67af0-8c30-449e-8e8b-ad334ec8d88c