Only create bookmark bar buttons that we need, not "for all". Create the "off the side" menu on demand (since that has "all the rest" in it).
BUG= http://crbug.com/32843

TEST=1) Do the "sync 1000+ bookmarks" test from the bug and watch Chrome NOT run out of memory.
Make sure these bookmarks are all children of the bar for it to be even harder.
2) Add about 5 bookmarks.  Make sure you can grow and shrink (in
width) the window to cause the off-the-side chevron to appear and
disappear as appropriate, and that the off-the-side menu always has
the right contents (bookmarks which fall off the side).
3) Change theme then relaunch; make sure theme set on launch for bookmarks.
4) Using both a big and a small width, make sure the chevron shows (or
not, as appropriate) on launch of Chrome.

Review URL: http://codereview.chromium.org/853001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@41359 4ff67af0-8c30-449e-8e8b-ad334ec8d88c