Add chromium-side support for history.{push,replace}State.
Enables the feature so that all but 2 of the related layout tests pass.

Modifies TestShell to correctly update its location bar as navigations occur.
It was incorrectly showing firstPartyForCookies for some crazy reason.

Modifies glue_serialize.cc to store the state object associated with a session
history entry.

Modifies navigation_controller.cc to always replace the current navigation
entry when observing an in-page navigation.  This is required since the page ID
isn't changing for an in-page navigation.

BUG=29393
R=brettw
TEST=covered by layout tests

Review URL: http://codereview.chromium.org/1036003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@41850 4ff67af0-8c30-449e-8e8b-ad334ec8d88c