2 experiments: DNS prefetch limit concurrency: TCP split a packet
Some firewalls apparently try to preclude a "syn flood to host" by limiting
the number of syn's (used to open a TCP/IP socket) that are outstanding
without having received a syn-ack.  Presumably this is to prevent a user
from participating in a syn-flood attack (which traditional sends a lot
of syn packets, with false return addresses, resulting in no responses).

Apparently this firewall technology has in some cases been extended
to include UDP sessions for which there has been no response, and this
may include DNS resolutions.  Since the prefetcher currently resolves
as many as 8 names simultaneously, this is remarkably close to the 
reported threshold of 10 un-answered connections.  This test attempts
to limit connections to 2, 4, or 6, so that we can see if this helps
users.

In TCP, the RTO remains (under windows) at a full 3 seconds until after the
first ack is received.  As a result, if the first data packet sent (after
the SYN) is lost, then TCP won't resend until after 3 seconds without an ack.
As a test, we split up the first packet into two parts (the second part
containing only one byte).  This is done as an A/B test, and we'll see
if we get a measurable improvement in page-load-time latency.

Finally, to get better page load stats, I adjusted the PLT histograms
so that we record a "final" time for abandoned pages when they are
closed (even if they didn't finish rendering, etc.).  This should give
a much more fair PLT comparison for all network latency experiments.


BUG=3041
BUG=12754
r=mbelshe,darin
Review URL: http://codereview.chromium.org/1088002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@42181 4ff67af0-8c30-449e-8e8b-ad334ec8d88c