Fix an exception thrown when pasting in BMM
Since the handle(e) function is called by a separate function instead of as an
event handler when pasting, it was trying to call preventDefault on a
non-existant event and throwing an exception.

Do we need to do something for paste to prevent the system beep? I'm not sure
how to cause the beep to test it out.

Review URL: http://codereview.chromium.org/1473001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@42869 4ff67af0-8c30-449e-8e8b-ad334ec8d88c