Hook up extension apps notification permission, take two
This is the chromium side of a change which will wait to land on the webkit side landing. (https://bugs.webkit.org/show_bug.cgi?id=36625)

It changes the NotificationPresenter to pass the sourceURL, rather than the SecurityOrigin in checking permission.

The full URL is required to match the app extent.

BUG=32361, 31024
TEST=NONE

Review URL: http://codereview.chromium.org/1383001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@43162 4ff67af0-8c30-449e-8e8b-ad334ec8d88c