REVISION BACKED OUT
Revert 43779 - linux: explicitly unhighlight buttons before disabling

If your mouse is hovering the back button when the button is disabled
(i.e. you click back and it takes some time for the navigate to go
through), when the button is reenabled GTK restores the highlighted
state to the button.

I reproduced this in a very simplified naive app on the bug and had
the same result, so I think this workaround is ok.

BUG=13729
TEST=do the actions listed in the bug

Review URL: http://codereview.chromium.org/1521016

TBR=evan@chromium.org
Review URL: http://codereview.chromium.org/1599021

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@43783 4ff67af0-8c30-449e-8e8b-ad334ec8d88c