Flush audio data after seek
There is about 400 to 600 ms of audio that needs to be flushed
after seek, this have a pretty UX effect. This is partially
fixed in this patch by reducing the amount of data down to
about at most 500 ms stored in the hardware buffer.

This patch clears data in the software buffer in the browser
process when seek happens. We could reduce the amount of
hardware buffer to further reduce the amount of lag but that
can be fixed by subsequent patch.

BUG=24150
TEST=audio still works, audio still play after seek
     the playback of old data after seek is substantially
     reduced.

Few changes in this patch:
1. Flush software buffer after seek in browser process
2. Get rid of prerolling, this actually has not effect at all, so getting rid of useless code

Needs to be done after this patch:
1. Further reduce the remaining data after seek and pause
2. Still hit the DCHECK in ClockImpl::Play(), this doesn't seem to be a new problem introduced in this patch

Review URL: http://codereview.chromium.org/1508021

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@43983 4ff67af0-8c30-449e-8e8b-ad334ec8d88c