Adding 2 preferences for configuring "NextEngine" and "Trigger" hot-keys.
This change also disables the latter hot-key so that users can't disable IBus. This is preferable since "current input method is xkb:us::eng (for example) and it is enabled" and "current one is xkb:us::eng but it's disabled" are almost indistinguishable.

This change depends on http://codereview.chromium.org/1528023 and http://codereview.chromium.org/1610013.

BUG=none
TEST=boot chrome os, activate English and Chinese, click omnibox, press lang menu, press English to enable "xkb:us::eng", press omnibox, press ctrl+space, verify that the menu shows "Chi" (not "Eng").

Review URL: http://codereview.chromium.org/1553018

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@44072 4ff67af0-8c30-449e-8e8b-ad334ec8d88c