NTP: Add work around for layout bug that some users are hitting.
In some rare cases the reported window.innerWidth and the media query reports different widths. This is most likely due to some timing issues. This work around keeps track of what layout mode we are using (wasSmallGrid) and when we get the data as well as on DOMContentLoaded we double check to ensure that the mode is correct and if not we do another layout. So in worst case we do one more layout in the case where the layout was wrong.


BUG=25329
TEST=I haven't been able to repro so when testing this I introduced the *wrong* case manually in code.

Review URL: http://codereview.chromium.org/1650008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@44590 4ff67af0-8c30-449e-8e8b-ad334ec8d88c