Refactor the split pane element into a splitter element.
The reason for the refactoring is so that the splitter can be used in cases where there are multiple columns.

This also adds a workaround for https://bugs.webkit.org/show_bug.cgi?id=32230 which causes the splitter to not work correctly when the page is zoomed.

BUG=42282, 42290
TEST=Open the bookmark manager. Zoom the page. Then drag the splitter.

Review URL: http://codereview.chromium.org/1763004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@45356 4ff67af0-8c30-449e-8e8b-ad334ec8d88c