GTK: Put options dialogs in a different window group.
GTK grabs work on a per window group basis, so when a modal dialog comes up
(which performs a grab) and the user then clicks on something in the main
browser window that spawns an InfoBubble, the InfoBubble's code that handles
hiding the currently grabbed widget will hide the modal dialog and it is lost
until restart.

Since we need to remove the centering in ContentSettingsWindowGtk for this to
work, also change the centering behaviour for FontsLanguagesWindowGtk, which
has bothered me for a while.

BUG=41815
TEST=none

Review URL: http://codereview.chromium.org/1727012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@45871 4ff67af0-8c30-449e-8e8b-ad334ec8d88c