Remove height 100% from NTP when in the app launcher.
This is done by using chrome://newtab/#app-launcher and then map the hash to an attribute which is later used in the CSS. I investigated several alternative methods but they all lead to either a higher memory footprint (due to two copies of the NTP source in memory) or slowdown of NTP. No caching of the resource.

During my refactoring I *accidentally* fixed 35783 so I'm keeping that in there.

BUG=35783
TEST=Launch chrome with a new profile with the app launcher turned on.

Review URL: http://codereview.chromium.org/1846003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@46266 4ff67af0-8c30-449e-8e8b-ad334ec8d88c