Fix a unit-test that was crashing on valgrind bot.
The problem is that ClientSocketPoolBaseHelper holds a pointer to a task, and tries to call a method on that pointer during destruction.
However, if destruction of ClientSocketPoolBaseHelper occurs during message loop destruction, then pending tasks are deleted, so this "backup_task" pointer that it accesses may be invalid.

I worked around the problem here by avoiding this destruction code-path in the unit-test. It would be good to make ClientSocketPoolBaseHelper able to handle this case though.

BUG=43291

Review URL: http://codereview.chromium.org/1920003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@46521 4ff67af0-8c30-449e-8e8b-ad334ec8d88c