gtk: fix setting minimum window size to not produce conflicting info
Before we could potentially set a minimum size greater than the
maximum size.  Now we take whichever is larger: the size requested
in the resources and the actual size of the content.  This allows
the resource size to grow the window larger without chopping off
the content if the content is larger than the specified resources size.

BUG=40580
TEST=verified it looked ok in en, pt, and de (all three cases are different)

Review URL: http://codereview.chromium.org/2034005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@46873 4ff67af0-8c30-449e-8e8b-ad334ec8d88c