Use realpath() to find the path to the extension unpack dir on posix systems.
Extensions are unpacked by a sandboxed utility process.  The sandbox forbids file access outside the directory the extension will be unpacked in.  If the path to that directory contains a symbolic link, then unpacking will fail because following the link will cause file system access outside the sandbox path.  Use realpath() to get a symlink free path to the directory where the extension will be unpacked.

A similar issue exists on windows, with junctions instead of symlinks.  This will be fixed in another change.

BUG=13044,35198
TEST=FileUtilTest.RealPath

Review URL: http://codereview.chromium.org/2001013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@47032 4ff67af0-8c30-449e-8e8b-ad334ec8d88c