GTK: Redo button tints in GTK+ mode.
Use separate tints for icons in the omnibox and the other icons like page/app menu. We don't pick the tinting color to be the skyline accent color, but instead create one based off the accent color and the contrast between the window background and the label color.

BUG=43853
TEST=1) Darklooks should have white icons (both the page icons and in the omnibox) 2) DarkRoom should have page/app menu icons that match the label color on the "Other Bookmarks" button while the icons in the omnibox which are on a white background should visually be more orange and contrast with the white. 3) Ambiance/Radiance should have icons that contrast with their background instead of fading into it. 4) New Wave should have dark rust orange instead of light neon orange (again, contrasting with the background).

Review URL: http://codereview.chromium.org/2266004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@48401 4ff67af0-8c30-449e-8e8b-ad334ec8d88c