Merge DidPaintBackingStoreRects and DidScrollBackingStoreRect into a single DidUpdateBackingStore method so that I can move the call to ScrollWindowEx closer to the RedrawWindow call.
This is done to minimize the opportunity for a v-sync of the
display happening between the ScrollWindowEx call and the
RedrawWindow call, which gives the appearance of tearing or
flicker in some cases.

RenderWidgetHostViewMac now just treats the scroll rect as an
additional paint rect.

Since RenderWidgetHostViewGtk was just repainting the scroll
rect, I took this opportunity to avoid double painting if a
paint rect happened to intersect the scroll rect.

R=brettw,rohitrao
BUG=none
TEST=none
Review URL: http://codereview.chromium.org/2733010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@49532 4ff67af0-8c30-449e-8e8b-ad334ec8d88c