Implements policies to enable or disable the five Privacy checkboxes found in the Options dialog, in the Under the Hood section.
xib changes: added an enabled binding to the policy-controlled checkboxes in the "Under the Hood" preference pane so that they can be disabled when their value is overridden by policy. The methods for the bindings are in the preference window controller.

BUG=44394
TEST=ConfigurationPolicyPrefStoreTest.* and ConfigurationPolicyProviderWinTest.*  Manual tests: Set the policies externally.  Check that the checkbox in Options/Under the Hood/Privacy/search suggest... are disabled.  Also check that the check box is set/unset as done by the policy.

Review URL: http://codereview.chromium.org/2746006
Patch from Jean-Luc Brouillet <jeanluc@google.com>.

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@49675 4ff67af0-8c30-449e-8e8b-ad334ec8d88c