Use GetSystemSalt from the cryptohome lib instead of reading salt off disk.
We used to read the system salt from a file on disk, which meant that
we had to do a bunch of dancing to do some work during login on the
FILE thread instead of the UI thread.  Now, we no longer have to,
because there is an API to request the system's salt value from
cryptohomed over dbus.  This CL is a precursor to some of the other
login/cryptohome changes that are coming.

BUG=chromium-os:323
TEST=unit tests, and loading up a chromium os device and logging in

Review URL: http://codereview.chromium.org/2820006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@49799 4ff67af0-8c30-449e-8e8b-ad334ec8d88c