TranslateInfoBarDelegate2 were leaked.
(Relanding, the changed unit-test was flaky.
Turns out comparing pointers in that context to see that an object changed is a bad idea as a new object might be at the exact same address as an old object.)

This revealed that the NormalTranslate unit-test was incorrect (it was not testing that changing the source or target language from the after translate infobar causes a translation). This also revealed that returning reference string in the delegate is not a brilliant idea as when we switch infobars, the new infobar might be using a string from the old infobar that might have been deleted.

BUG=None
TEST=Linux heap check bot should be green.

Review URL: http://codereview.chromium.org/2878005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@50920 4ff67af0-8c30-449e-8e8b-ad334ec8d88c