This change allows the syncer thread to be paused and resumed when the syncer thread has not yet been started or it has been started and is waiting for a connection.
From an API standpoint, the caller of SyncerThread::Pause and SyncerThread::Resume should not need to consider if the thread is already running.  The channel responses (PAUSED and RESUMED) should be the same regardless.  Furthermore, edge cases like "pause before thread is started and resume after thread is started" should be handled correctly.

This change is supposed to be fixing various offline startup problems we've been having and it includes a new integration test for this.  However, it is not yet complete but we might need to check this is before it is ready.

TEST=unit
BUG=45814

Review URL: http://codereview.chromium.org/2814014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@51158 4ff67af0-8c30-449e-8e8b-ad334ec8d88c