Autocomplete and AutoFill entries are shown combined in the suggestions popup.
These changes merge popup menu suggestions for Autocomplete and AutoFill.  Suggestions are collected from both sources and returned to the renderer as a single list.  This combined list is given to WebKit to display in the suggestions popup.  When the user selects an item in the popup the appropriate result gets filled in the fields.

BUG=41328
TEST=AutoFillManagerTest.*

Review URL: http://codereview.chromium.org/2878020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@51736 4ff67af0-8c30-449e-8e8b-ad334ec8d88c