[Mac] Reset to Defaults should reset the startup pages list.
The URLs were actually getting reset, the UI was not updating. The original
issue was that mutable list Preference values could not be observed for
changes. That was worked-around on Mac, but when the broadcast notifications
were implemented (fixing the original issue), the Mac-specific work around was
incompatible and caused an infinte loop. This removes that work-around and
uses the cross-platform functionality.

BUG=48083
TEST=Chromium-->Preferences-->Basics-->Open the following pages. Add a page. Go to Under the Hood and Reset to Defaults. Go to Basics; no pages are in the list.

Review URL: http://codereview.chromium.org/2870048

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@51842 4ff67af0-8c30-449e-8e8b-ad334ec8d88c