AutoFill credit cards should be encrypted on the Mac
These changes add encryption support on Mac for the Encryptor class.  AES 128 bit is used for the encryption, and the auto-generated password is stored now in the Mac Keychain.  This implies the Encryptor class on Mac can now block for user input, and can fail if access is denied.

BUG=42038, 49131
TEST=EncryptorTest.CypherTextDiffers, EncryptorTest.DecryptError, EncryptorPasswordTest.*

Review URL: http://codereview.chromium.org/2943014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@52590 4ff67af0-8c30-449e-8e8b-ad334ec8d88c