Attempt 2 at landing this CL. The earlier one broke the build. We don't need to include the external_tab_container_win.h file in tab_contents. That is the only change.
Links opening up in the host browser in ChromeFrame would not show up as
visited. This was because top level link navigations
are directly sent via the OpenURL mechanism to the host browser and as a result
the visited link history is not updated.

Fix is to update the navigation history in the
ExternalTabContainer::OpenURLFromTab function. We currently use the
UpdateHistoryForNavigation function in TabContents which is also used by regular
navigations initiated within the renderer.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=37262

Bug=37262
TBR=amit

Review URL: http://codereview.chromium.org/2809067

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@53283 4ff67af0-8c30-449e-8e8b-ad334ec8d88c