[Gtk] Display the current content setting applying to a domain along with the domain.
This patch only affectes the gtk ui. Here is what it looks like: http://imgur.com/g6Gox

If third party cookies are blocked, a warning is shown. Next to each domain, the content setting is displayed that applies (or nothing if the default applies). When the user selects some domains and presses an button, the labels are updated accordingly.

BUG=49826
TEST=unit_tests

Review URL: http://codereview.chromium.org/3096001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@56315 4ff67af0-8c30-449e-8e8b-ad334ec8d88c