Disable outdated plug-ins on a per-plugin basis.
Change the automatic disabling of outdated plugins (with --disable-outdated-plugins) so that a single plugin is disabled if it is out of date instead of the whole group.

This makes it possible to fall back on a newer version of a plugin that would otherwise have lower priority.

BUG=47731,51728
TEST=Install an old version and a new version of a plugin, run with --disable-outdated-plugins. The old version should be disabled, and the effective version of the plugin group should be the new one.

Review URL: http://codereview.chromium.org/3117009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@56494 4ff67af0-8c30-449e-8e8b-ad334ec8d88c