Extract connection logic from HttpNetworkTransaction into a new HttpStreamFactory.  The HttpNetworkTransaction now deals exclusively with streams rather than connections directly.  This cut the size of HTN roughly in in half.
The HttpNetworkTransaction is still responsible for all proxy and server
authentication functions.  This is because the streams may come and go - we
could create a stream, have the server declare auth is needed, and then the
next attempt would be on a different stream.  So Auth belongs on the HNT.
The HNT no longer has direct access to the connection itself; instead, it
only knows of an HttpStream.

The StreamRequest, however, is responsible for determining whether the
connection needs to use a proxy, whether AlternateProtocols are available, and
whether the connection should be SPDY or HTTP.

Other changes:
  - moved some static configuration methods from HNT to HttpStreamFactory.
  - added some methods to the HttpStream.


BUG=none
TEST=all



Review URL: http://codereview.chromium.org/3171002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@56646 4ff67af0-8c30-449e-8e8b-ad334ec8d88c