Give keys friendly names in NSS and OS X
This will make them browsing them easier in Keychain Access and certutil.
It also gives them a more useful name in Keychain permission dialogs.

The friendly name is currently just the hostname which requested
the key. It would be nice to have some surrounding text with
localization, but this works for now.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/2806045

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@56774 4ff67af0-8c30-449e-8e8b-ad334ec8d88c