Remove dependency on X509Certificate::Cache behaviour when selecting a client certificate on Win
On OS X and Linux (cocoa/gtk), the original X509Certificate pointer supplied in SSLCertRequestInfo's client_certs is the one returned when a user selects a certificate.

On Windows, a new X509Certificate is created from the selected certificate and returned, rather than the original X509Certificate. This translates to a dependency on X509Certificate::Cache to return the same certificate, which, while presently is true, is an implementation specific detail that should not be relied upon.

BUG=none
TEST=SSL client authentication continues to work on Windows

Review URL: http://codereview.chromium.org/3170019

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@57003 4ff67af0-8c30-449e-8e8b-ad334ec8d88c