[Chrome OS] Wire up ownership API from libcros
I added an asynchronous API to libcros for setting the Owner's key
on Chrome OS.  This CL wires that out through LoginLibrary and also
refactors the OwnerManager class into the OwnershipService, a
Singleton that will provide access to the entire Ownership API --
which includes methods for whitelisting users and signed preferences.
The OwnerManager class is now just there to manage the Owner's
key and operations with that key.

A consumer of this API can call methods of OwnershipService, which will post tasks to run methods from OwnerManager (on the FILE thread if appropriate).  The OwnerManager will, in turn, call methods of LoginLibrary to perform owner key operations.  Since these calls to LoginLibrary are asyncrhonous, OwnerManager implements LoginLibrary::Delegate and passes |this|, so that LoginLibrary can call back to the OwnerManager when asynchronous operations are completed.

BUG=chromium-os:4488
TEST=unit tests

Review URL: http://codereview.chromium.org/3141031

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@57269 4ff67af0-8c30-449e-8e8b-ad334ec8d88c