chromeos: Remove accelerator from fullscreen-exit text.
This currently displays a nonsensical message, but I can't
think of any good way to change the text to describe the
key, e.g. "Exit full screen (fourth key from the left on the
top row of the keyboard)". :-P

We'd ideally include an image of the keycap in the middle of
the text, but that seems like it'll require a decent amount
of Views hacking, and will be a pain to maintain while we're
still in the process of playing around with things like font
sizes (but I've created an image that works with the current
text size, if someone more Views-proficient than me wants to
take this on).  Glen agreed that this is an acceptable
compromise in the meantime.

BUG=chromium-os:3916
TEST=tried it

Review URL: http://codereview.chromium.org/3181042

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@57449 4ff67af0-8c30-449e-8e8b-ad334ec8d88c