Re-land 57460 - Add hidden component app for web store.
Fixed ExtensionStartupTest which was essentially counting the number of component extensions. Now it doesn't.

Original review here: http://codereview.chromium.org/3116040

This patch adds a component which does not appear on the NTP, but does enforce that gallery urls are isolated in their own process. This is in anticipation of exposing extension api bindings to the gallery so it can install/detect/uninstall extensions & apps.

Note that this patch works correctly with the --apps-gallery-url. If specified, the replacement url will be inserted into the component app's extent

BUG=27431
TBR=aa

Review URL: http://codereview.chromium.org/3104039

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@57520 4ff67af0-8c30-449e-8e8b-ad334ec8d88c