Move the SingletonSocket to a temporary directory
This is to workaround problems on certain network filesystems (notably AFS)
which do not support Unix domain sockets. We move the sockets into a temporary
folder and symlink. To avoid the possibility of a dangling link to a missing
(and later intercepted) remote directory, we create and check cookie files
and rely on the stickiness of /tmp/ to avoid a race condition in the check.

R=mattm
BUG=44606
TEST=ProcessSingletonLinuxTest.*

Review URL: http://codereview.chromium.org/2838034

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@57623 4ff67af0-8c30-449e-8e8b-ad334ec8d88c