Fix safe-browsing prefix comparator to sort stably.
SafeBrowsingDatabaseBloom::PairCompare() compared SBPairs.  The prefix
part is a 32-bit hash which should be uniformly distributed.  The
previous code used a trick of subtracting the prefixes and returning
the result.  If arg1<arg2, that resulted in a negative return value,
if arg1>arg2, positive, if ==, zero.  Since the distribution was
uniform, 1/4 of prefix comparisons gave incorrect results, which
caused qsort() and bsearch() gave incorrect results.

In practice, this caused some add prefixes to remain in the database
when their opposite sub prefixes were present.  Most of these would be
gone after a few updates (as items were added or removed, the sort
within a chunk would change, allowing more matches to occur).  The
excess add prefixes may have resulted in excess queries up to the
servers, but should not have resulted in false misses (sites escaping
the filter).

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/3284011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@58172 4ff67af0-8c30-449e-8e8b-ad334ec8d88c