Fix Firefox data import for Mac.
r58258 broke Firefox import on the Mac by trying to use a value from g_browser_process, which is not accessible in Mac out-of-process import.

Because this value isn't used or needed for the Mac build, don't even create it in the Mac version.

BUG=54265
TEST=Firefox import works on Mac again.

Review URL: http://codereview.chromium.org/3341016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@58775 4ff67af0-8c30-449e-8e8b-ad334ec8d88c