Change UI font at startup time if necessary, based on the UI locale.
As of writing, ChromeDroidSans is used as the UI font (see
/etc/gtk-2.0/gtkrc). This setting does not work well for Japanese users
as Chinese glyphs are used for Kanji characters. For Japanese UI, we
should use a Japanese font.

Introduce IDS_UI_FONT_FAMILY_CROS to control the UI font based on the
UI locale. For now, app_locale_settings_ja.xtb only contains a
non-default value: IPAPGothic.

Just for the record, I originally tried to use "DroidSans Japanese"
but for some reason, this caused space characters (0x20) to be garbled
in the wrench menu. Hence, IPAPGotchic is chosen for now.

TEST=manually on the netbook, and ubuntu
BUG=chromium-os:4038

Review URL: http://codereview.chromium.org/3275008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@58913 4ff67af0-8c30-449e-8e8b-ad334ec8d88c