2nd attempt at http://codereview.chromium.org/3305003/show
The difference between this patch and the other one is plumbing setup_in_test_mode through the Initialization process, rather than using a separate SetupInTestMode() method, which happens too late to safely stop the syncer thread.  [see syncapi.cc|h for most of the changes.]

---

This patch removes: authenticator.cc, auth_watcher.cc

removes calls to user_settings.cc, removes an authenticate PB request to the server, and moves token storage into the Chrome TokenService. This patch introduces the SigninManager, which is an interim solution for user management prior to moving the system into chrome.

Other changes include removing the dependency on the sync backend to be running while the sync wizard is intially displayed. This means that the backend can be brought up in response to credentials becoming available. The backend now is always provided credentials on startup. If an auth error occurs, it propogates it up via a notification. Some event handlers were removed and streamlined for more straightforward sync system startup.

BUG=51001, 50293, 35158
TEST=Unit tests && Start up sync, log in, log out, run with expired credentials, run with new gaia credentials, run with gaia credentials updated while system is syncing. Try logging in with incorrect username. Trigger CAPTCHA. Try logging out and in repeatedly. Check about:sync works. Try going offline and back online again. Expire gaia credentials and try renewing it with the UI dialog.
Review URL: http://codereview.chromium.org/3342025

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@58993 4ff67af0-8c30-449e-8e8b-ad334ec8d88c