Added full support for invoking before unload and unload handlers on ChromeFrame rendered pages. This allows a webpage to put up a confirmation dialog in its beforeunload handler and potentially cancel the operation. We only support invoking unload handlers on the page for IE full tab mode. To achieve this the active document handles the OLECMDID_ONUNLOAD exec command which is passed by the DOCHOST to the object which allows us to potentially cancel the operation.
Thanks to Stoyan for his help in authoring parts of this CL.

The AutomationMsg_RunUnloadHandlers message which is used only by ChromeFrame is now a sync
message which returns back a bool indicating whether the unload operation can be continued
or not. The ExternalTabContainer now implements the BeforeUnloadFired method in the
TabContentsDelegate and aborts the unload operation if the user chose to not proceed with
the unload.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=33200

Bug=33200
Test=Covered by existing unload event test. Will add a test which validates whether a page
can cancel the unload operation in a subsequent CL.

Review URL: http://codereview.chromium.org/3450014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@59854 4ff67af0-8c30-449e-8e8b-ad334ec8d88c