Autofill profile pop-up should trigger for mouse double-clicks
Removes the cancellation logic for separated AutoFill and Autocomplete.  This is no longer needed.  The recent refactoring of the popup handling causes this logic to always hit the "cancel" scenario and thus return only Autocomplete suggestions and never AutoFill suggestions.  But in fact, we get the Autocomplete-only results in the correct cases by virtue of the field type queried.

BUG=54875
TEST=AutoFillManagerTest.GetFieldSuggestions*

Review URL: http://codereview.chromium.org/3405015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@59950 4ff67af0-8c30-449e-8e8b-ad334ec8d88c