[Chrome OS] Infrastucture for doing offline/online login simultaneously
These are meant to wrap and replace large chunks of code in GoogleAuthenticator,
once I rewrite that code to do offline and online login at the same time.
That's why there's some code that's still in GoogleAuthenticator that's
refactored into some of these classes.

Essentially, one creates an AuthAttemptState object, which will only allow
its mutable contents to be read/written on the IO thread.  Then, you pass it
to an OnlineAttempt which, when done, stores results in the AuthAttemptState.
You can do the same thing with the various CryptohomeOp classes.  They call out
over DBus to asynchronously do stuff with the user's cryptohome and then,
when the callbacks come in, these object post a call over to the IO thread to
update the contents of the AuthAttemptState instance they're holding on to.

BUG=chromium-os:4929
TEST=unit testss

Review URL: http://codereview.chromium.org/3407008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@60190 4ff67af0-8c30-449e-8e8b-ad334ec8d88c