Mark a failing test as FLAKY and added a potential fix + log messages.
The test was failing on mac due to fractional numbers and this CL fixes it.

The test also fails on windows sometimes complaining about a missing comma,
so I have added a log message and marked the test as flaky. I'll be watching
the tree for failures and log messages

BUG=51337
TEST=Tree turns green
TBR=bulach@chromium.org

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@60927 4ff67af0-8c30-449e-8e8b-ad334ec8d88c