linux: Behave predictably on ctrl+backspace
The default behaviour of GtkTextView on ctrl+backspace is a little counter
intuitive. If there is selected text and the cursor is on the left edge of
the selection, then instead of removing the selected text, it removes the
word prior to the selection. So in cases where ctrl+backspace is pressed
and text is selected, treat it as a normal backspace.

BUG=18224

TEST=Type in 'goo' in the omnibox. That should autocomplete to 'google.com'.
Pressing ctrl+backspace should remove the suggested 'gle.com'.

Review URL: http://codereview.chromium.org/3446021

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@61132 4ff67af0-8c30-449e-8e8b-ad334ec8d88c