A subtle bug was exposed when the following was added to Preferences::RegisterUserPrefs:
  prefs->RegisterIntegerPref(prefs::kLabsTalkEnabled, 0);

Because prefs::kLabsTalkEnabled = "extensions.settings.ggnioahjipcehijkhpdjekioddnjoben.state",
"extensions.settings" in the DEFAULT pref store is forced to a DictionaryValue
instead of an empty value so that data can be stored in it. This invalidaded a
test in PrefService::GetMutableDictionary, causing GetMutableDictionary to
return a pointer to a DictionaryValue in the default pref store instead of
creating a dictionary in the user pref store.

BUG=http://code.google.com/p/chromium-os/issues/detail?id=7066
TEST=See bug; should no longer repro. Also, test preferences in general.

Review URL: http://codereview.chromium.org/3533008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@61209 4ff67af0-8c30-449e-8e8b-ad334ec8d88c