Policy to disable printing
Make Chrome refuse to print if the preference kPrintingEnabled is set to false. In this case, the 'Print...' option in the wrench menu is disabled, and if printing is invoked from JavaScript, then Chrome reports that no printer is installed.

Changes to this preference after startup are also respected.

BUG=54380
TEST=manual

Review URL: http://codereview.chromium.org/3357011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@61222 4ff67af0-8c30-449e-8e8b-ad334ec8d88c