DOMUI Options: Fix vertical alignment of radio and checkboxes
This also makes the label text line up correctly if the label needs to wrap.

This also fixes a bug where the menu could be scrolled off the screen.

I went through and fixed a bunch of indentation errors in the HTML I touched.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/3590006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@61544 4ff67af0-8c30-449e-8e8b-ad334ec8d88c