ExtensionApiTest.Tabs: fix window type, re-enable test
When you created a window with type='popup', the resulting window.type
was incorrectly 'app'.  After discussion, we concluded that a popup
app window should be called a popup.

This was a real regression that was masked by the test being marked flaky
for unrelated reasons.  :(

(While I'm at it, fixed the ordering of some assertEq's to make the error
messages more understandable.)

BUG=48920
TEST=ExtensionApiTest.Tabs

Review URL: http://codereview.chromium.org/3594014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@61722 4ff67af0-8c30-449e-8e8b-ad334ec8d88c