Fix TemplateURLFetcher test.
Address the issue of relying on files not available to all build servers which
caused some valgrind bots to hang.

Fixed a leak by changing the order of shutdown of items to ensure that the I/O
thread was still running when it was needed to delete an object. This resulted
in a lot of code re-arrangement because threads need to be deleted in a certain
order as well. It seemed less fragile overall to concentrate this complexity in
one place more than it was.

Also this addresses Pawel feedback from http://codereview.chromium.org/3548017

BUG=58242,58243
TEST=unit_tests --gtest_filter=Sea*:Temp*

Review URL: http://codereview.chromium.org/3614011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@61837 4ff67af0-8c30-449e-8e8b-ad334ec8d88c