Add timeout to clearing server data
Background: the sync nudge infrastructure is highly tuned for sync scenarios -- a fire-and-forget nudge is sent tothe sync thread which kicks off sync.  If the syncer is disconnected, throttled, paused, etc, then the nudge is dropped as a sync cannot happen.

Problem: one-shot commands such as clearing server data can also be dropped without any acknowledgment, which means that the UI is never updated.  There were a few alternatives for fixing this: 1) threading the clear through, 2) adding a timeout, or 3) fixing the nudge infrastructure to support more reliable messaging semantics.  #3 was the ideal choice, however not doable for M8.  Until the syncer is redesigned, #2 is our stop-gap solution.

On success, sync is disabled and the dialog dialog is closed.  We do not care if a timeout event occurred thereafter, and theres no reason to show an error message to the user at that point.  A similar situation exists if you are looking at the dialog and someone clears from another browser: the dialog closes itself without warning and without regard to any local requests to clear.

BUG=57360
TEST=

Review URL: http://codereview.chromium.org/3720001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@62424 4ff67af0-8c30-449e-8e8b-ad334ec8d88c