Fix a failing test for the extension webstorePrivate API.
The test has been failing very frequently, always hitting the expectation
in FakeProfileSyncService::RemoveObserver that the observer being removed isn't
the currently registered one.

I think the problem is that I was waiting until the PromptBrowserLoginFunction
destructor (which is called by the ref count going to 0) to have the function
remove itself as a listener, when really once the OnStateChanged function 
fires and sync setup is complete we don't want to be listening anymore.

BUG=59419
TEST=The ExtensionWebstorePrivateBrowserTest.BrowserLogin browser test should
stop failling

Review URL: http://codereview.chromium.org/3974002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@63315 4ff67af0-8c30-449e-8e8b-ad334ec8d88c